Style changes as per comments and added implementation note explaining the distributed approach.

+++ mllib/src/main/scala/org/apache/spark/mllib/stat/test/KSTest.scala
@@ -22,13 +22,26 @@ import org.apache.commons.math3.stat.inference.KolmogorovSmirnovTest
 
 import org.apache.spark.rdd.RDD
 
-
 /**
  * Conduct the two-sided Kolmogorov Smirnov test for data sampled from a
  * continuous distribution. By comparing the largest difference between the empirical cumulative
  * distribution of the sample data and the theoretical distribution we can provide a test for the
  * the null hypothesis that the sample data comes from that theoretical distribution.
  * For more information on KS Test: https://en.wikipedia.org/wiki/Kolmogorov%E2%80%93Smirnov_test
+ *
+ * Implementation note: We seek to implement the KS test with a minimal number of distributed
+ * passes. We sort the RDD, and then perform the following operations on a per-partition basis:
+ * calculate an empirical cumulative distribution value for each observation, and a theoretical
+ * cumulative distribution value. We know the latter to be correct, while the former will be off by
+ * a constant (how large the constant is depends on how many values precede it in other partitions).
+ * However, given that this constant simply shifts the ECDF upwards, but doesn't change its shape,
+ * and furthermore, that constant is the same within a given partition, we can pick 2 values
+ * in each partition that can potentially resolve to the largest global distance. Namely, we
+ * pick the minimum distance and the maximum distance. Additionally, we keep track of how many
+ * elements are in each partition. Once these three values have been returned for every partition,
+ * we can collect and operate locally. Locally, we can now adjust each distance by the appropriate
+ * constant (the cumulative sum of # of elements in the prior partitions divided by the data set
+ * size). Finally, we take the maximum absolute value, and this is the statistic.
  */
 private[stat] object KSTest {
 
@@ -46,13 +59,12 @@ private[stat] object KSTest {
    */
   def testOneSample(data: RDD[Double], cdf: Double => Double): KSTestResult = {
     val n = data.count().toDouble
-    val localData = data.sortBy(x => x).mapPartitions {
-      part =>
-        val partDiffs = oneSampleDifferences(part, n, cdf) // local distances
-        searchOneSampleCandidates(partDiffs) // candidates: local extrema
-        }.collect()
+    val localData = data.sortBy(x => x).mapPartitions { part =>
+      val partDiffs = oneSampleDifferences(part, n, cdf) // local distances
+      searchOneSampleCandidates(partDiffs) // candidates: local extrema
+      }.collect()
     val ksStat = searchOneSampleStatistic(localData, n) // result: global extreme
-    evalOneSampleP(ksStat, n.toInt)
+    evalOneSampleP(ksStat, n.toLong)
   }
 
   /**
@@ -63,13 +75,12 @@ private[stat] object KSTest {
    */
   def testOneSample(data: RDD[Double], createDist: () => RealDistribution): KSTestResult = {
     val n = data.count().toDouble
-    val localData = data.sortBy(x => x).mapPartitions {
-      part =>
-        val partDiffs = oneSampleDifferences(part, n, createDist) // local distances
-        searchOneSampleCandidates(partDiffs) // candidates: local extrema
-    }.collect()
+    val localData = data.sortBy(x => x).mapPartitions { part =>
+      val partDiffs = oneSampleDifferences(part, n, createDist) // local distances
+      searchOneSampleCandidates(partDiffs) // candidates: local extrema
+      }.collect()
     val ksStat = searchOneSampleStatistic(localData, n) // result: global extreme
-    evalOneSampleP(ksStat, n.toInt)
+    evalOneSampleP(ksStat, n.toLong)
   }
 
   /**
@@ -89,19 +100,20 @@ private[stat] object KSTest {
     : Iterator[Double] = {
     // zip data with index (within that partition)
     // calculate local (unadjusted) ECDF and subtract CDF
-    partData.zipWithIndex.map {
-      case (v, ix) =>
-        // dp and dl are later adjusted by constant, when global info is available
-        val dp = (ix + 1) / n
-        val dl = ix / n
-        val cdfVal = cdf(v)
-        // if dp > cdfVal the adjusted dp is still above cdfVal, if dp < cdfVal
-        // we want negative distance so that constant adjusted gives correct distance
-        if (dp > cdfVal) dp - cdfVal else dl - cdfVal
-    }
+    partData.zipWithIndex.map { case (v, ix) =>
+      // dp and dl are later adjusted by constant, when global info is available
+      val dp = (ix + 1) / n
+      val dl = ix / n
+      val cdfVal = cdf(v)
+      // if dp > cdfVal the adjusted dp is still above cdfVal, if dp < cdfVal
+      // we want negative distance so that constant adjusted gives correct distance
+      if (dp > cdfVal) dp - cdfVal else dl - cdfVal
+      }
   }
 
-  private def oneSampleDifferences(partData: Iterator[Double], n: Double,
+  private def oneSampleDifferences(
+      partData: Iterator[Double],
+      n: Double,
       createDist: () => RealDistribution)
     : Iterator[Double] = {
     val dist = createDist()
@@ -119,10 +131,9 @@ private[stat] object KSTest {
   private def searchOneSampleCandidates(partDiffs: Iterator[Double])
     : Iterator[(Double, Double, Double)] = {
     val initAcc = (Double.MaxValue, Double.MinValue, 0.0)
-    val partResults = partDiffs.foldLeft(initAcc) {
-      case ((pMin, pMax, pCt), currDiff) =>
-        (Math.min(pMin, currDiff), Math.max(pMax, currDiff), pCt + 1)
-    }
+    val partResults = partDiffs.foldLeft(initAcc) { case ((pMin, pMax, pCt), currDiff) =>
+      (Math.min(pMin, currDiff), Math.max(pMax, currDiff), pCt + 1)
+      }
     Array(partResults).iterator
   }
 
@@ -140,18 +151,17 @@ private[stat] object KSTest {
     val initAcc = (Double.MinValue, 0.0)
     // adjust differences based on the # of elements preceding it, which should provide
     // the correct distance between ECDF and CDF
-    val results = localData.foldLeft(initAcc) {
-      case ((prevMax, prevCt), (minCand, maxCand, ct)) =>
-        val adjConst = prevCt / n
-        val pdist1 = minCand + adjConst
-        val pdist2 = maxCand + adjConst
-        // adjust by 1 / N if pre-constant the value is less than cdf and post-constant
-        // it is greater than or equal to the cdf
-        val dist1 = if (pdist1 >= 0 && minCand < 0) pdist1 + 1 / n else Math.abs(pdist1)
-        val dist2 = if (pdist2 >= 0 && maxCand < 0) pdist2 + 1 / n else Math.abs(pdist2)
-        val maxVal = Array(prevMax, dist1, dist2).max
-        (maxVal, prevCt + ct)
-    }
+    val results = localData.foldLeft(initAcc) { case ((prevMax, prevCt), (minCand, maxCand, ct)) =>
+      val adjConst = prevCt / n
+      val pdist1 = minCand + adjConst
+      val pdist2 = maxCand + adjConst
+      // adjust by 1 / N if pre-constant the value is less than cdf and post-constant
+      // it is greater than or equal to the cdf
+      val dist1 = if (pdist1 >= 0 && minCand < 0) pdist1 + 1 / n else Math.abs(pdist1)
+      val dist2 = if (pdist2 >= 0 && maxCand < 0) pdist2 + 1 / n else Math.abs(pdist2)
+      val maxVal = Array(prevMax, dist1, dist2).max
+      (maxVal, prevCt + ct)
+      }
     results._1
   }
